support drawing lines

+++ image.h
@@ -24,13 +24,18 @@ struct Image {
 void image_clear(struct Image *image);
 void image_draw_line(struct Image *image, struct Line *line);
 static inline enum Color image_get(struct Image *image, uint8_t x, uint8_t y) {
-    uint8_t index = (x >> 2) * IMAGE_WIDTH + y; 
-    uint8_t byte = image->buffer[index];
-    switch (x & 0b11) {
-        case 0: return (byte & 0b00000011);
-        case 1: return (byte & 0b00001100) >> 2;
-        case 2: return (byte & 0b00110000) >> 4;
-        case 3: return (byte & 0b11000000) >> 6;
-        default: return WHITE;
-    }
+  uint16_t index = (x >> 2) * IMAGE_WIDTH + y;
+  uint8_t byte = image->buffer[index];
+  switch (x & 0b11) {
+  case 0:
+    return (byte & 0b00000011);
+  case 1:
+    return (byte & 0b00001100) >> 2;
+  case 2:
+    return (byte & 0b00110000) >> 4;
+  case 3:
+    return (byte & 0b11000000) >> 6;
+  default:
+    return WHITE;
+  }
 }
